requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12        # Required for Form + File upload in FastAPI

# ── LangGraph & LangChain ─────────────────────────────────────────────────
langgraph==0.2.55
langchain==0.3.9
langchain-core==0.3.21
langchain-groq==0.2.1

# ── LLM provider ──────────────────────────────────────────────────────────
groq==0.11.0
httpx==0.27.2                   # Pinned to fix groq 'proxies' kwarg crash

# ── PDF processing ────────────────────────────────────────────────────────
PyMuPDF==1.24.13                # fitz — fast, accurate PDF text + table extraction

# ── Data validation ───────────────────────────────────────────────────────
pydantic==2.10.2
pydantic-settings==2.6.1

# ── Utilities ─────────────────────────────────────────────────────────────
python-dotenv==1.0.1            # Load .env for local dev
tenacity==9.0.0                 # Retry utilities (used by langchain internally)

# ── Dev / Testing (optional, comment out in prod) ─────────────────────────
# pytest==8.3.3
# pytest-asyncio==0.24.0
# httpx==0.27.2                 # TestClient for FastAPI